requirements.txt
requests
pandas
sqlalchemy
streamlit
python-dotenv
pymysql